// apps/cms/src/payload-types.ts
/* tslint:disable */
/**
 * This file was automatically generated by Payload CMS.
 * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface Config {}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string
  email?: string
  resetPasswordToken?: string
  resetPasswordExpiration?: string
  loginAttempts?: number
  lockUntil?: string
  role: 'admin' | 'developer' | 'editor' | 'viewer'
  createdAt: string
  updatedAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "services".
 */
export interface Service {
  id: string
  enableAPIKey?: boolean
  apiKey?: string
  apiKeyIndex?: string
  email?: string
  resetPasswordToken?: string
  resetPasswordExpiration?: string
  loginAttempts?: number
  lockUntil?: string
  name: string
  createdAt: string
  updatedAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string
  url?: string
  filename?: string
  mimeType?: string
  filesize?: number
  width?: number
  height?: number
  sizes: {
    thumbnail: {
      url?: string
      width?: number
      height?: number
      mimeType?: string
      filesize?: number
      filename?: string
    }
    card: {
      url?: string
      width?: number
      height?: number
      mimeType?: string
      filesize?: number
      filename?: string
    }
    tablet: {
      url?: string
      width?: number
      height?: number
      mimeType?: string
      filesize?: number
      filename?: string
    }
  }
  alt?: string
  createdAt: string
  updatedAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "rewards".
 */
export interface Reward {
  id: string
  _status?: 'draft' | 'published'
  name: string
  image?: string | Media
  huddleId:
    | 'public'
    | 'c52f2155-154d-4425-be36-9807f14f7dc0'
    | 'd4ff07db-512c-4a2a-bf3f-adc5973e67b0'
    | 'a83e0e80-5f78-42a2-8fa7-a61a08734dba'
    | 'b7742841-79b3-4206-ba6e-0a6a0c99969e'
  code: string
  dateComingSoon: string
  dateSelecting: string
  dateAwarded: string
  claimBy: string
  legacyYear?: number
  awardedTo?: string
  createdAt: string
  updatedAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "polls".
 */
export interface Poll {
  id: string
  _status?: 'draft' | 'published'
  title: string
  slug: string
  opensAt: string
  expiresAt: string
  huddleId:
    | 'public'
    | 'c52f2155-154d-4425-be36-9807f14f7dc0'
    | 'd4ff07db-512c-4a2a-bf3f-adc5973e67b0'
    | 'a83e0e80-5f78-42a2-8fa7-a61a08734dba'
    | 'b7742841-79b3-4206-ba6e-0a6a0c99969e'
  image?: string | Media
  question: string
  options: {
    text: string
    image?: string | Media
    id?: string
  }[]
  blindVote: boolean
  public: boolean
  minSelections: number
  maxSelections: number
  category?: 'Collectibles' | 'Experiences' | 'Moments' | 'Platform'
  text: {
    [k: string]: unknown
  }[]
  detailedDescription: {
    [k: string]: unknown
  }[]
  votingInstructions?: string
  views?: number
  reactions: {
    emoji: string
    count?: number
    id?: string
  }[]
  createdAt: string
  updatedAt: string
}